opt: move rule-specific functions to their own custom funcs files
Previously, most custom functions for normalization rules were held
in the custom_funcs file, which meant it had >2000 lines of code. This
made it difficult to find functions for reuse.

This patch moves functions which are only or nearly only used in a single
normalization rule to the file corresponding to that rule. General-purpose
functions which are likely to be reused are left in custom_funcs, which
has been renamed to general_funcs. The hope is that this refactoring will
make reusable functions easier to find.

To make backports easier, these changes will be backported to previous
releases from 20.2.

Release note: None
