+++ pkg/sql/opt/norm/project_funcs.go
@@ -9,3 +9,87 @@
 // licenses/APL.txt.
 
 package norm
+
+import (
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
+	"github.com/cockroachdb/cockroach/pkg/sql/types"
+)
+
+// CanMergeProjections returns true if the outer Projections operator never
+// references any of the inner Projections columns. If true, then the outer does
+// not depend on the inner, and the two can be merged into a single set.
+func (c *CustomFuncs) CanMergeProjections(outer, inner memo.ProjectionsExpr) bool {
+	innerCols := c.ProjectionCols(inner)
+	for i := range outer {
+		if outer[i].ScalarProps(c.mem).OuterCols.Intersects(innerCols) {
+			return false
+		}
+	}
+	return true
+}
+
+// MergeProjections concatenates the synthesized columns from the outer
+// Projections operator, and the synthesized columns from the inner Projections
+// operator that are passed through by the outer. Note that the outer
+// synthesized columns must never contain references to the inner synthesized
+// columns; this can be verified by first calling CanMergeProjections.
+func (c *CustomFuncs) MergeProjections(
+	outer, inner memo.ProjectionsExpr, passthrough opt.ColSet,
+) memo.ProjectionsExpr {
+	// No need to recompute properties on the new projections, since they should
+	// still be valid.
+	newProjections := make(memo.ProjectionsExpr, len(outer), len(outer)+len(inner))
+	copy(newProjections, outer)
+	for i := range inner {
+		item := &inner[i]
+		if passthrough.Contains(item.Col) {
+			newProjections = append(newProjections, *item)
+		}
+	}
+	return newProjections
+}
+
+// MergeProjectWithValues merges a Project operator with its input Values
+// operator. This is only possible in certain circumstances, which are described
+// in the MergeProjectWithValues rule comment.
+//
+// Values columns that are part of the Project passthrough columns are retained
+// in the final Values operator, and Project synthesized columns are added to
+// it. Any unreferenced Values columns are discarded. For example:
+//
+//   SELECT column1, 3 FROM (VALUES (1, 2))
+//   =>
+//   (VALUES (1, 3))
+//
+func (c *CustomFuncs) MergeProjectWithValues(
+	projections memo.ProjectionsExpr, passthrough opt.ColSet, input memo.RelExpr,
+) memo.RelExpr {
+	newExprs := make(memo.ScalarListExpr, 0, len(projections)+passthrough.Len())
+	newTypes := make([]types.T, 0, len(newExprs))
+	newCols := make(opt.ColList, 0, len(newExprs))
+
+	values := input.(*memo.ValuesExpr)
+	tuple := values.Rows[0].(*memo.TupleExpr)
+	for i, colID := range values.Cols {
+		if passthrough.Contains(colID) {
+			newExprs = append(newExprs, tuple.Elems[i])
+			newTypes = append(newTypes, *tuple.Elems[i].DataType())
+			newCols = append(newCols, colID)
+		}
+	}
+
+	for i := range projections {
+		item := &projections[i]
+		newExprs = append(newExprs, item.Element)
+		newTypes = append(newTypes, *item.Element.DataType())
+		newCols = append(newCols, item.Col)
+	}
+
+	tupleTyp := types.MakeTuple(newTypes)
+	rows := memo.ScalarListExpr{c.f.ConstructTuple(newExprs, tupleTyp)}
+	return c.f.ConstructValues(rows, &memo.ValuesPrivate{
+		Cols: newCols,
+		ID:   values.ID,
+	})
+}
